Made first version of smarter line wrie function
I am building on a linewirteanimate function wich wil wrtie a line to
canvas element like if it is an animation. Fist it could only write a
line to left under side. I am now building the fucntion so it can wirte
horizontal. The final plan is to make it so mart it can detect how to
write and let it write horizontal, vertical and corner to corner.